Agregadas funciones para buscar en Provider
