+++ public/manifest.json
@@ -8,14 +8,8 @@
       "type": "image/x-icon"
     },
     {
-      "src": "logo192.png",
       "type": "image/png",
       "sizes": "192x192"
-    },
-    {
-      "src": "logo512.png",
-      "type": "image/png",
-      "sizes": "512x512"
     }
   ],
   "start_url": ".",
