+++ src/App.js
@@ -1,7 +1,7 @@
-import {React, useState} from 'react';
+import {React} from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { PokemonProvider } from './contexts/PokemonProvider';
 import './App.css';
-import { FavouriteProvider } from './contexts/favouriteContext';
 import Pokedex from './pages/Pokedex/Pokedex';
 import PokedexList from './pages/PokedexList/PokedexList';
 import PokemonDetail from './pages/PokemonDetail/PokemonDetail';
@@ -12,14 +12,8 @@ import "./styles/styles.scss"
 
 export default function App() {
   
-  const [favourites, setFavourites] = useState([]);
-  
-  const updateFavoritePokemons = (name) => {
-    console.log(name);
-  };
-
   return (
-    <FavouriteProvider>
+    <PokemonProvider>
       <BrowserRouter>
         <Navbar />     
         <Routes>       
@@ -29,7 +23,7 @@ export default function App() {
           <Route element={<PokemonDetail/>} path="/details" />
         </Routes>
       </BrowserRouter>
-    </FavouriteProvider>
+    </PokemonProvider>
   );
 }
