+++ src/contexts/PokemonContext.jsx
@@ -1,0 +1,2 @@
+import { createContext } from "react";
+export const PokemonContext = createContext()
