+++ src/styles/base/_settings.scss
@@ -12,17 +12,14 @@
 	outline: none;
 	border-radius: 25px;
 	padding: 10px 20px;
-	
 	font-family: inherit;	
 	letter-spacing: 2px;
 	font-weight: bold;
-	cursor: pointer;
-		
+	cursor: pointer;		
 	background-color: $color-3;
 	color: #fff;
 }
 	
-
 // Class-color Pokemon
     $color-fire: #ff7402;
 	$color-grass: #9bcc50;
