+++ src/styles/components/_Navbar.scss
@@ -3,6 +3,34 @@ header{
     align-items: center;
     justify-content: space-between;
     padding: 40px 80px;
+    form{
+        display: flex;
+        align-items: center;
+        gap: 15px;
+        .form-group{
+            display: flex;
+            align-items: center;
+            gap: 10px;
+            border: 1px solid #555;
+            padding: 15px 20px;
+            border-radius: 25px;
+            .icon-search{
+                width: 20px;
+                height: 20px;
+                stroke: #555;
+            }
+            input{
+                font-family: inherit;
+                width: 300px;
+                border: none;
+                outline: none;
+                font-size: 15px;
+            }
+        }
+        .btn-search{
+            @include button-1()
+        }
+    }
 
     img{
         width: 150px;
