+++ src/styles/pages/_Pokedex.scss
@@ -4,34 +4,6 @@ main{
     .coso{
         display: flex;
         justify-content: space-around;
-        form{
-            display: flex;
-            align-items: center;
-            gap: 15px;
-            .form-group{
-                display: flex;
-                align-items: center;
-                gap: 10px;
-                border: 1px solid #555;
-                padding: 15px 20px;
-                border-radius: 25px;
-                .icon-search{
-                    width: 20px;
-                    height: 20px;
-                    stroke: #555;
-                }
-                input{
-                    font-family: inherit;
-                    width: 300px;
-                    border: none;
-                    outline: none;
-                    font-size: 15px;
-                }
-            }
-            .btn-search{
-                @include button-1()
-            }
-        }
     }
     .pokedex-grid{
         display: flex;
